% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneConversions.R
\name{any2entrez}
\alias{any2entrez}
\title{Convert Symbols or Ensembl gene ids to Entrez Ids}
\usage{
any2entrez(gene_ids, verbose = TRUE)
}
\arguments{
\item{gene_ids}{a character vector of gene ids from the Hugo gene sybmol or Ensembl annotation, respectively}

\item{verbose}{logical whether to message progress reports.}
}
\value{
a character vector of gene symbols, beware of NAs
}
\description{
Two reasons for NA return: the external gene id could not be found in the reference (org.Hs.Db)
or there is no Entrez Id for that particular symbol or Ensembl Id, respectively.
}
