% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneConversions.R
\name{any2symbol}
\alias{any2symbol}
\title{Convert Entrez or Ensembl gene ids to gene symbols}
\usage{
any2symbol(gene_ids, verbose = TRUE)
}
\arguments{
\item{gene_ids}{a character vector of gene ids from the Ensembl or Entrez annotation, respectively}

\item{verbose}{logical whether to message progress reports.}
}
\value{
a character vector of gene symbols, beware of NAs
}
\description{
Two reasons for NA return: the external gene id could not be found in the reference (org.Hs.Db)
or there is no gene symbol for that particular external gene id.
}
