% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper-helpers.R
\name{aREA_single}
\alias{aREA_single}
\title{aREA for a single regulon}
\usage{
aREA_single(ges, regulon, wm = NULL)
}
\arguments{
\item{ges}{named numeric vector representing a gene expression signature}

\item{regulon}{a single regulon of a given TF}

\item{wm}{optional weight matrix for the gene expression signature}
}
\value{
a list with enrichment score, normalized enrichment score and p-value
}
\description{
This function performs aREA enrichment analysis for a single regulon and returns
a normalized enrichment score and corresponding p-value
}
