% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{basic_rank_norm}
\alias{basic_rank_norm}
\title{Basic rank normalization of a gene expression matrix}
\usage{
basic_rank_norm(counts, tpm = FALSE, pmOne = TRUE,
  gausstrans = FALSE, ...)
}
\arguments{
\item{counts}{Matrix containing raw read counts (generally).}

\item{tpm}{logical, adjust for gene length before rank transformation}

\item{pmOne}{logical, whether to rescale values so they are distributed between -1 and +1}

\item{gausstrans}{logical, whether to rescale the values to quantiles of the gaussian distribution}

\item{...}{further arguments passed to the tpm function}
}
\value{
basic rank normalized matrix.
}
\description{
Basic rank normalization of a gene expression matrix
}
