% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{basic_signature}
\alias{basic_signature}
\title{basic_signature}
\usage{
basic_signature(eset, tpm = FALSE, ...)
}
\arguments{
\item{eset}{Numeric matrix of gene expression (raw counts or pre-processed)}

\item{tpm}{logical, whether tpm normalization should be carried out (use if eset is in raw counts)}

\item{...}{further arguments to tpm function}
}
\value{
Numeric matrix with a basic signature for each sample
}
\description{
This function prepares a gene expression matrix before metaVIPER. After rank
normalization of each sample, the row medians are subtracted and every column
is divided by the row median absolute deviation (MAD). It will filter out genes with a mad
of zero.
}
