% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{brks_heatmap}
\alias{brks_heatmap}
\title{Center heatmap color gradient at zero}
\usage{
brks_heatmap(emat, color_palette)
}
\arguments{
\item{emat}{numeric matrix}

\item{color_palette}{vector representing colors, no specific type needed, only used for length}
}
\value{
a numeric vector specifying breaks for a heatmap (pheatmap package)
}
\description{
Little helper function to center a color palette around 0 for heatmaps depicting relative signatures such
as log2 fold changes or Z-scores. Mostly tailored towards the pheatmap R package.
}
