% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{collapse_multi}
\alias{collapse_multi}
\title{Collapsing many-to-one relationships for gene expression matrices}
\usage{
collapse_multi(expmat, fac, method = c("mean", "median", "sum"),
  verbose = TRUE)
}
\arguments{
\item{expmat}{matrix whose rows can partly be summarized using a factor}

\item{fac}{factor used for summarizing rows}

\item{method}{method of summary, defaults to mean}

\item{verbose}{whether to inform user regarding progress}
}
\value{
new matrix with rows or columns in order of the reference list
}
\description{
Collapsing many-to-one relationships for gene expression matrices
}
