% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixOp.R
\name{colVar}
\alias{colVar}
\title{colVar}
\usage{
colVar(A, removeNA = TRUE)
}
\arguments{
\item{A}{a numeric matrix}

\item{removeNA}{logical should NA values be removed from colSums/colMeans ?}
}
\value{
vector of variance per column
}
\description{
This function replaces the apply(A, 2, var) call
}
