% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{DEtransform}
\alias{DEtransform}
\title{Variance stabilization transformation for RNAseq data}
\usage{
DEtransform(emat, fitType = c("parametric", "local"), noise = TRUE, seed = 42)
}
\arguments{
\item{emat}{matrix containing the raw counts, with genes in rows and samples in columns}

\item{fitType}{Character string indicating the type of fit for the dispersion (see DESeq::estimateDispersions)}

\item{noise}{Logical, indicating whether to add a shot of random noise to the data (splitting ties)}

\item{seed}{Integer indicating the fixed seed for random numbers, 0 for not setting the seed}
}
\value{
Expression matrix
}
\description{
This function stabilizes the variance, transform the data and add shot noise to the data
}
