% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{extract_features}
\alias{extract_features}
\title{Extract rows or columns from a matrix in order of interest}
\usage{
extract_features(expmat, reflist, margin = 1)
}
\arguments{
\item{expmat}{matrix with row and column names}

\item{reflist}{character vector of row or column names to be extracted}

\item{margin}{where to look for the reflist, by default rows}
}
\value{
new matrix with rows or columns in order of the reference list
}
\description{
Extract rows or columns from a matrix in order of interest
}
