% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper-helpers.R
\name{geneSets2regulon}
\alias{geneSets2regulon}
\title{Convert regular (one-tailed) gene sets to regulons}
\usage{
geneSets2regulon(geneSets)
}
\arguments{
\item{geneSets}{list of named lists representing gene sets of gene symbols}
}
\value{
regulon containing one-tailed gene sets as regulators
}
\description{
So they can be used with the viper function
}
