% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{gs_ovlp}
\alias{gs_ovlp}
\title{Score character vectors of genes and their overlap with certain gene sets}
\usage{
gs_ovlp(genelist, genesets)
}
\arguments{
\item{genelist}{character vector of gene identifiers}

\item{genesets}{named list of character vectors}
}
\value{
a tibble of genes and gene sets they appear in
}
\description{
Score character vectors of genes and their overlap with certain gene sets
}
