% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{gsea1-class}
\alias{gsea1-class}
\title{The gsea1 class}
\description{
This class contains the results generated by the gsea1T function
}
\section{Slots}{

   \describe{
     \item{\code{ES}:}{numeric value containing the enrichment score}
     \item{\code{RS}:}{numeric vector containing the running sum}
     \item{\code{signature}:}{named numeric vector containg the gene expression signature}
     \item{\code{es_idx}:}{integer indicating the position of the enrichment score}
     \item{\code{gs_idx}:}{integers indicating the position the genes from the gene set within the signature}
     \item{\code{ledge}:}{character vector of gene names occuring before the ES in the signature}
     \item{\code{ledge_index}:}{integer vector of leading edge gene positions}
     \item{\code{NES}:}{numeric value containing the enrichment score, will only be filled by gsea_null}
     \item{\code{pval}:}{numeric value containing the p-value, will only be filled by gsea_null}
     \item{\code{null_es}:}{numeric vector of enrichment scores from a permutation based nullmodel}
     \item{\code{mean_nulles}:}{numeric value containing the mean of the relevant tail of null enrichment scores}
     }
}

