% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_Subramanian.R
\name{gsea1T}
\alias{gsea1T}
\title{This function performs one-tailed Gene Set Enrichment Analysis as described by Subramanian et al.}
\usage{
gsea1T(
  signature,
  gS,
  weight = 1,
  sorting = c("decreasing", "increasing"),
  onlyES = FALSE
)
}
\arguments{
\item{signature}{Named vector of gene-level statistics (preferably t-statistics)}

\item{gS}{Gene set of interest}

\item{sorting}{character indicating how to sort the signature (default is decreasing)}

\item{onlyES}{logical, whether only the enrichment score should be returned}

\item{w}{Weight for each gene-level statistic (default = 1)}
}
\value{
A list of class "gsea1" containing:
\describe{
\item{ES}{enrichment score}
\item{RS}{running sum}
\item{signature}{input signature}
\item{es_idx}{index of the maximum enrichment score}
\item{gs_idx}{indices of the genes in the gene set in the signature}
\item{ledge}{gene identifiers for the genes in the leading edge}
\item{ledge_index}{indices of the leading edge genes in the signature}
}
}
\description{
This function performs one-tailed Gene Set Enrichment Analysis as described by Subramanian et al.
}
