% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{gsea2-class}
\alias{gsea2-class}
\title{The gsea2 class}
\description{
This class contains the results generated by the gsea2T and the gsea_regulon function, respectively.
}
\section{Slots}{

   \describe{
     \item{\code{signature}:}{named numeric vector containg the gene expression signature}
     \item{\code{ES_pos}:}{numeric value containing the enrichment score of gene set 1 (pos. targets)}
     \item{\code{ES_neg}:}{numeric value containing the enrichment score of gene set 2 (neg. targets)}
     \item{\code{ES_pos_idx}:}{integer value indicating the position of the ES_pos}
     \item{\code{ES_neg_idx}:}{integer value indicating the position of the ES_neg}
     \item{\code{RS_pos}:}{numeric vector containing the running sum of gene set 1}
     \item{\code{RS_neg}:}{numeric vector containing the running sum of gene set 2}
     \item{\code{gs_idx_pos}:}{integers indicating the position the genes from the gene set 1 within the signature}
     \item{\code{gs_idx_neg}:}{integers indicating the position the genes from the gene set 2 within the signature}
     \item{\code{ledge_pos}:}{character vector of gene names from gene set 1 occuring before the ES in the signature}
     \item{\code{ledge_neg}:}{character vector of gene names from gene set 2 occuring before the ES in the signature}
     \item{\code{ledge_index_pos}:}{integer vector of leading edge gene positions from gene set 1}
     \item{\code{ledge_index_neg}:}{integer vector of leading edge gene positions from gene set 2}
     \item{\code{null_es_pos}:}{numeric vector of enrichment scores from a permutation based nullmodel for gene set 1}
     \item{\code{null_es_neg}:}{numeric vector of enrichment scores from a permutation based nullmodel for gene set 2}
     \item{\code{NES_pos}:}{numeric value containing the normalized enrichment score of gene set 1, will only be filled by gsea_null}
     \item{\code{NES_neg}:}{numeric value containing the normalized enrichment score of gene set 2, will only be filled by gsea_null}
     \item{\code{pval_pos}:}{numeric value containing the p-value of gene set 1, will only be filled by gsea_null}
     \item{\code{pval_neg}:}{numeric value containing the p-value of gene set 2, will only be filled by gsea_null}
     }
}

