% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_Lim.R
\name{gsea2ES}
\alias{gsea2ES}
\title{ES for 2 tail GSEA (Lim et al., 2009) - two character vectors}
\usage{
gsea2ES(signature, set1, set2, weight = 1)
}
\arguments{
\item{signature}{Named vector of gene-level statistics (preferably t-statistics)}

\item{set1}{character vector, i.e. the first gene set}

\item{set2}{character vector, i.e. the second gene set}

\item{weight}{Weight for each gene-level statistic (default = 1)}
}
\value{
An overall enrichment score for two complementary gene sets on a signature
}
\description{
This computes one overall enrichment score for two complementary gene sets on a given signature
as described in Lim et al., Pac. Symp. Biocomput. 2009, 504-515.
}
