% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_Subramanian.R
\name{gsea2T}
\alias{gsea2T}
\title{This function performs GSEA of two gene sets for a given signature}
\usage{
gsea2T(
  signature,
  set1,
  set2,
  sorting = c("decreasing", "increasing"),
  weight = 1
)
}
\arguments{
\item{signature}{Named vector of gene-level statistics (preferably t-statistics)}

\item{set1}{character vector, i.e. the first gene set}

\item{set2}{character vector, i.e. the second gene set}

\item{sorting}{character indicating how to sort the signature (default is decreasing). Passed to gsea1T.}

\item{w}{weight for each gene-level statistic (default = 1)}
}
\value{
a list object containing for each gene set:
\describe{
\item{signature}{input signature}
\item{ES_pos}{enrichment score set1}
\item{ES_neg}{enrichment score set2}
\item{ES_pos_idx}{enrichment score set1 index}
\item{ES_neg_idx}{enrichment score set2 index}
\item{RS_pos}{running sum for set1}
\item{RS_neg}{running sum for set2}
\item{gs_idx_pos}{indices of the set1 genes in the signature}
\item{gs_idx_neg}{indices of the set2 genes in the signature}
\item{ledge_pos}{gene identifiers of the leading edge for set1}
\item{ledge_neg}{gene identifiers of the leading edge for set2}
\item{ledge_idx_pos}{indices of the leading edge gene set members - set1}
\item{ledge_idx_neg}{indices of the leading edge gene set members - set2}
}
}
\description{
Similar to \code{gsea_regulon} but without any information on Mode of Regulation
}
