% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_null.R
\name{gsea2T_null}
\alias{gsea2T_null}
\title{Calculating a null distribution of enrichment scores for two-tail GSEA}
\usage{
gsea2T_null(signature, set1, set2, w = 1, perm = 1000, seed = 42,
  plot_hist = FALSE, verbose = TRUE)
}
\arguments{
\item{signature}{named numeric vector of gene-level statistics}

\item{set1}{character vector of gene names of positive targets}

\item{set2}{character vector of gene names of negative targets}

\item{w}{numeric indicating the weights used for GSEA (defaults to 1)}

\item{perm}{integer indicating the number of permutations to carry out (defaults to 1000)}

\item{seed}{integer for random number generation during the permutation process}

\item{plot_hist}{logical on whether to plot a histogram of the permutation-based enrichment scores}

\item{verbose}{logical on whether to message progress report}
}
\value{
A gsea_null object
}
\description{
This function generates a GSEA null distribution of enrichment scores based on gene shuffling, i.e.
the positions of genes in the signature are permuted rather than phenotype labels.
}
