% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_Subramanian.R
\name{gsea_regulon}
\alias{gsea_regulon}
\title{This function performs GSEA for the negative and positive targets, respectively, of a regulatory gene.}
\usage{
gsea_regulon(signature, regulon, sorting = c("decreasing", "increasing"),
  weight = 1)
}
\arguments{
\item{signature}{Named vector of gene-level statistics (preferably t-statistics)}

\item{regulon}{list object as returned by aracne2regulon (viper package)
\describe{
\item{tfmode}{named numeric vector of the mode of regulation (MOR) values}
\item{likelihood}{numeric vector of interaction confidence values (based on mutual information)}
}}

\item{sorting}{character indicating how to sort the signature (default is decreasing). Passed to gsea1T.}

\item{w}{weight for each gene-level statistic (default = 1)}
}
\value{
a list object containing for each gene set:
\describe{
\item{signature}{input signature}
\item{ES_pos}{enrichment score positive targets}
\item{ES_neg}{enrichment score negative targets}
\item{ES_pos_idx}{enrichment score positive targets index}
\item{ES_neg_idx}{enrichment score negative targets index}
\item{RS_pos}{running sum for positive targets}
\item{RS_neg}{running sum for negative targets}
\item{gs_idx_pos}{indices of the positive targets in the signature}
\item{gs_idx_neg}{indices of the negative targets in the signature}
\item{ledge_pos}{gene identifiers of the leading edge for the positive targets}
\item{ledge_neg}{gene identifiers of the leading edge for the negative targets}
\item{ledge_idx_pos}{indices of the leading edge gene set members - positive targets}
\item{ledge_idx_neg}{indices of the leading edge gene set members - negative targets}
}
}
\description{
This function performs GSEA for the negative and positive targets, respectively, of a regulatory gene.
}
