% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{hallmarkFounders}
\alias{hallmarkFounders}
\title{Information on source gene sets of MSigDb Hallmark collection
See Liberzon, Cell Syst, 2015 for details. PMID: 26771021}
\format{
list of lists: for 50 Hallmark gene sets, one list each comprising all founders
}
\description{
Information on source gene sets of MSigDb Hallmark collection
See Liberzon, Cell Syst, 2015 for details. PMID: 26771021
}
\author{
HCM
}
\keyword{data}
