% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{homologyInfo}
\alias{homologyInfo}
\title{Human and Mouse Homology data
Information on 16469 genes for which there is exactly a 1:1 mapping from a mouse gene id to a human gene id.}
\format{List with two elements containing gene \strong{symbol} and \strong{entrez id} data, respectively
\describe{
\item{Symbol}{named character vector of mouse gene symbols, where names represent human symbols}
\item{Entrez}{named character vector of mouse gene entrez ids, where names represent human entrez ids}
}}
\source{
\url{http://www.informatics.jax.org/homology.shtml}
}
\description{
Human and Mouse Homology data
Information on 16469 genes for which there is exactly a 1:1 mapping from a mouse gene id to a human gene id.
}
\author{
HCM
}
\keyword{data}
