% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{kaboxplot}
\alias{kaboxplot}
\title{Plot quantiles per column for a numeric matrix}
\usage{
kaboxplot(eset, quants = c(0.05, 0.25, 0.5, 0.75, 0.95),
  brewer_name = "Dark2", ylab = "Normalized expression",
  xlab = "Samples", ...)
}
\arguments{
\item{eset}{numeric matrix typically containing gene expression data}

\item{quants}{integer of number of top MR to consider}

\item{brewer_name}{RColorBrewer palette, defaults to Dark2}

\item{...}{further arguments passed to matplot}
}
\value{
graph with the respective quantiles across the sample space
}
\description{
Plot quantiles per column for a numeric matrix
}
