% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lapply_pair}
\alias{lapply_pair}
\title{lapply function to apply functions in a pair-wise manner between all distinct elements of a list}
\usage{
lapply_pair(x, skip_self = TRUE, f, ...)
}
\arguments{
\item{x}{list}

\item{skip_self}{logical indicating on whether function should be applied to the same element of the list, may be set to FALSE for symmetry reasons}

\item{f}{the function to be applied between each pair of elements of the list. Function needs to take two arguments: i and j.}
}
\value{
list object with the results
}
\description{
lapply function to apply functions in a pair-wise manner between all distinct elements of a list
}
