% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{ligandsReceptors}
\alias{ligandsReceptors}
\title{Information on ligands and receptors
As derived from Ramilowski et al., Nature Communications, 2016. PMID: 26198319.}
\format{
tibble with information on ligands and their receptors
}
\description{
Information on ligands and receptors
As derived from Ramilowski et al., Nature Communications, 2016. PMID: 26198319.
}
\author{
HCM
}
\keyword{data}
