% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lrcor}
\alias{lrcor}
\title{For two matched matrices, compute correlation (Spearman) between two related genes with the first
coming from matrix A and the second coming from matrix B.}
\usage{
lrcor(mat1, mat2, lrlist)
}
\arguments{
\item{mat1}{numeric matrix}

\item{mat2}{numeric matrix}

\item{lrlist}{named list containing one or multiple strings}
}
\value{
a tibble with genes from mat1 matched to those from mat2 and their spearman correlation
}
\description{
For two matched matrices, compute correlation (Spearman) between two related genes with the first
coming from matrix A and the second coming from matrix B.
}
