% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneConversions.R
\name{mouse2human}
\alias{mouse2human}
\title{Convert mouse gene symbols or entrez ids to human counterparts}
\usage{
mouse2human(eset, type = c("Symbol", "Entrez"))
}
\arguments{
\item{eset}{gene expression matrix with murine expression data}

\item{type}{Character to indicate the gene identifier, defaults to gene symbol}
}
\value{
Numeric matrix with human gene identifiers of the same type
}
\description{
Convert mouse gene symbols or entrez ids to human counterparts
}
