% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{msigdb}
\alias{msigdb}
\title{Modules c1 through c6 and hallmark gene sets from the MSigDb v6.2
Downloaded from the Broad Institute on October 2018}
\format{
List with 12 elements containing human gene symbols
\describe{
\item{c1.all}{hallmark gene sets, n = 326}
\item{c2.cgp}{canonical pathways, n = 3433}
\item{c2.cp}{transcription factor targets, n = 1329}
\item{c3.mir}{micro RNA targets, n = 221}
\item{c3.tft}{transcription factor targets, n = 615}
\item{c4.cgn}{cancer gene neighborhoods, n = 427}
\item{c4.cm}{cancer modules factor, n = 431}
\item{c5.bp}{GO biological processes, n = 4436}
\item{c5.cc}{GO cellular components, n = 580}
\item{c5.mf}{GO molecular function, n = 901}
\item{c6.all}{oncogenic signatures, n = 189}
\item{h.all}{hallmark gene sets, n = 50}
}
}
\description{
Modules c1 through c6 and hallmark gene sets from the MSigDb v6.2
Downloaded from the Broad Institute on October 2018
}
\author{
HCM
}
\keyword{data}
