% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_env.R
\name{mypar}
\alias{mypar}
\title{Quick changes to the plotting environment}
\usage{
mypar(a = 1, b = 1, brewer_n = 9, brewer_name = "Set1", ...)
}
\arguments{
\item{a}{number of rows (passed to mfrow)}

\item{b}{number of columns (passed to mfrow)}

\item{brewer_n}{number of colours to retrieve from RColorBrewer palette}

\item{brewer_name}{RColorBrewer palette}

\item{...}{further arguments passed to par}
}
\value{
By default only a new colour palette
}
\description{
Quick changes to the plotting environment
}
