% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{order_heatmap}
\alias{order_heatmap}
\title{Order the columns of a heatmap within the constraints of a factor
Generally, a left to right gradient will be achieved where column means will
increase from left to right.}
\usage{
order_heatmap(expmat, factorCol, factorRow = NULL, rev = FALSE)
}
\arguments{
\item{expmat}{expression matrix}

\item{factorCol}{factor with metadata information for each column in the expression matrix}

\item{factorRow}{factor with metadata information for each row in the expression matrix}
}
\value{
ordered expression matrix ready for heatmaps (do not cluster columns or rows, respectively)
}
\description{
Order the columns of a heatmap within the constraints of a factor
Generally, a left to right gradient will be achieved where column means will
increase from left to right.
}
