% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{p2z}
\alias{p2z}
\title{Convert p-values from a two tailed test to quantiles of the normal distribution}
\usage{
p2z(pvals, stats)
}
\arguments{
\item{pvals}{numeric vector of raw p-values}

\item{stats}{numeric vector of test statistics, e.g. t-statistics (for the sign)}
}
\value{
numeric vector of Z-scores
}
\description{
Convert p-values from a two tailed test to quantiles of the normal distribution
}
