% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{pdaClassifiers}
\alias{pdaClassifiers}
\title{Gene expression signatures used to classify PDA.
Derived from Collisson et al., Moffitt et al. and Bailey et al.}
\description{
Gene expression signatures used to classify PDA.
Derived from Collisson et al., Moffitt et al. and Bailey et al.
}
\author{
HCM
}
\keyword{data}
