% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pick_probes}
\alias{pick_probes}
\title{Find best probe if there are multiple per gene}
\usage{
pick_probes(expmat, fac, method = c("IQR", "mad", "mean", "median"))
}
\arguments{
\item{expmat}{matrix whose rows can partly be summarized using a factor}

\item{fac}{factor used for summarizing rows}

\item{method}{method of summary, defaults to IQR}
}
\value{
new matrix where rownames correspond to factor level names (e.g. gene symbols)
}
\description{
It will return the probe with the highest summary measure (IQR, mean, etc.)
}
