% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA_Subramanian.R
\name{plot.gsea1}
\alias{plot.gsea1}
\title{Plot results from a one-tailed GSEA}
\usage{
\method{plot}{gsea1}(gsea1, color = "cornflowerblue", main = "",
  plotSignature = FALSE, signatureNames = NULL, signatureType = NULL,
  ledge = TRUE, fixY = NULL, ...)
}
\arguments{
\item{gsea1}{gsea1 object as returned from \code{gsea1T}}

\item{color}{character or integer indicating color choice (defaults to cornflowerblue)}

\item{main}{Character vector given as argument for convenience}

\item{plotSignature}{Logical whether to include a graph on the signature (e.g. t-statistics)}

\item{signatureNames}{Character vector indicating the conditions being compared. Defaults to NULL. Otherwise a charater vector of length two.}

\item{signatureType}{Character to indicate the type of gene level statistic (e.g. t-statistic).}

\item{ledge}{logical whether to plot the leading edge line}

\item{fixY}{numeric vector specyfying y-axis limits, defaults to NULL}

\item{...}{Given for compatibility to the plot generic function}
}
\value{
Nothing, a plot is generated in the default output device
}
\description{
This function generates a plot
}
