% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_signatures_plot.R
\name{plot_fgseaRes}
\alias{plot_fgseaRes}
\title{Plot one-tailed GSEA on one signature for multiple gene sets}
\usage{
plot_fgseaRes(
  ges,
  geneSets,
  fgseaRes,
  ledge_only = FALSE,
  signatureNames = NULL,
  color = c("firebrick2", "royalblue"),
  strip_PWnames = TRUE,
  ...
)
}
\arguments{
\item{ges}{gene expression signature, a named vector of Z-scores, t-statistics, or log2 fold changes}

\item{geneSets}{a named list of character vectors representing the gene sets}

\item{fgseaRes}{Output of fgsea, will be used to get NES and adjusted p values}

\item{ledge_only}{Logical, whether to include all genes of gene set or only those in leading edge}

\item{signatureNames}{Character vector of length 2, specifying the experimental conditions}

\item{color}{Vector of two components indicating the colors for each part of the signature}

\item{strip_PWnames}{Logical, whether to remove commonly used Pathway prefixes, e.g. HALLMARK}

\item{...}{adjustment of cex for pathway names}
}
\value{
Nothing, a plot is generated in the default output device
}
\description{
This function generates a plot of the member genes of a given gene set on a given signature
for multiple gene sets.
}
