% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_signatures_plot.R
\name{plot_OneGs_MultSig}
\alias{plot_OneGs_MultSig}
\title{Plot ranks of a given gene set across several gene expression profiles from a numeric matrix.}
\usage{
plot_OneGs_MultSig(sigmat, geneset, color = "black", gs_label = NULL, ...)
}
\arguments{
\item{sigmat}{gene expression matrix, usually of raw values such as read counts, TPM or FPKM}

\item{geneset}{a character vector representing one gene set}

\item{color}{single character string indicating a color for the bar lines}

\item{gs_label}{single character string to label the gene set under investigation}

\item{...}{given for compatibility, particularly for the adjustment of cex for various text labels}
}
\value{
Nothing, a plot is generated in the default output device
}
\description{
This function generates a plot of the member genes of a given gene set across multiple signatures.
Its first intention is to illustrate the so called 'percent rank' of a given gene from a given
gene set.
}
\examples{
sigmat <- matrix(rnorm(1e5), ncol = 10, nrow = 1e4)
rownames(sigmat) <- paste0('Gene', seq(nrow(sigmat)))
colnames(sigmat) <- paste0('Sample', seq(ncol(sigmat)))
set.seed(42)
gs <- sample(rownames(sigmat), size = 100)
plot_OneGs_MultSig(sigmat = sigmat, geneset = gs)
}
