% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_signatures_plot.R
\name{plot_OneReg_MultSig}
\alias{plot_OneReg_MultSig}
\title{Plot two tails of one regulon on multiple signatures}
\usage{
plot_OneReg_MultSig(
  sigmat,
  tf,
  regulon,
  color = c("salmon", "cornflowerblue"),
  nes = NULL,
  padj = NULL,
  ...
)
}
\arguments{
\item{sigmat}{a matrix with genes in rows and signatures in columns. Typically a Z-score, t-statistic of log2 fold change matrix}

\item{tf}{a character vector of a regulatory protein, used to subset regulon object}

\item{regulon}{named regulon object similar to output of viper::aracne2regulon}

\item{color}{Vector of two components indicating the colors for the negative and positive parts of the regulon}

\item{nes}{numeric vector of normalized enrichment scores, must match number of columns in sigmat}

\item{padj}{numeric vector of adjusted p-values, must match number of columns in sigmat}
}
\value{
Nothing, a plot is generated in the default output device
}
\description{
This function generates a plot
}
