% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PurIST.R
\name{purIST}
\alias{purIST}
\title{PurIST}
\usage{
purIST(emat, raw = FALSE)
}
\arguments{
\item{emat}{expression matrix with row (=gene) and column (=sample) names. Genes must be supplied as HUGO gene symbols}

\item{raw}{logical indicating on whether to return a matrix with raw expression values for all pairs for inspection (default = FALSE)}
}
\value{
a named numeric vector of probabilities reflecting class membership or a matrix of pasted expression values
}
\description{
This function implements single-sample classification for PDAC specimen based on their raw gene expression profiles.
Idealy you would normalize for gene length although it does work with microarray intensities as well, so should be fine
with normalized counts just accounting for different library sizes.
}
\details{
PMID: 31754050

Probability key:
< 0.1: strongly favors classical
0.1 - 0.4: likely classical
0.4 - 0.5: lean classical
0.5 - 0.6: lean basal-like
0.6 - 0.9: likely basal-like
>0.9: strongly favors basal-like
}
