% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{randomize_eset}
\alias{randomize_eset}
\title{randomize_eset}
\usage{
randomize_eset(eset, seed = 1, margin = 2, ...)
}
\arguments{
\item{eset}{Numeric matrix with row and column names}

\item{seed}{integer, defaults to 1}

\item{margin}{where to randomize, defaults to columns}

\item{...}{Further arguments passed to apply}
}
\value{
new matrix with randomized columns or rows, respectively
}
\description{
randomize columns or rows, respectively of a matrix
}
