% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper-helpers.R
\name{randomize_regulon}
\alias{randomize_regulon}
\title{Generate a randomized network}
\usage{
randomize_regulon(regulon, universe = NULL, seed = 42)
}
\arguments{
\item{regulon}{regulon object - see output of viper::aracne2regulon}

\item{universe}{set to NULL, can be supplied to provide the gene universe to chose targets from}

\item{seed}{before random sampling}
}
\value{
regulon with randomized targets and preserved topology
}
\description{
The procedure applies the Maslov-Sneppen procedure, i.e. substitutes regulatory protein targets
with the same number of randomly assigned targets, thus producing a random program topology
while preserving the number of target genes of each RP.
}
