% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{read_matrix}
\alias{read_matrix}
\title{Tidyverse way to read in gene expression matrices}
\usage{
read_matrix(df)
}
\arguments{
\item{df}{path to a tab-separated dataframe whose first column includes gene identifiers}
}
\value{
Data matrix ready to play
}
\description{
Tidyverse way to read in gene expression matrices
}
