% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper-helpers.R
\name{reg2tibble}
\alias{reg2tibble}
\title{transform a regulon object to a data frame / tibble}
\usage{
reg2tibble(reg, annotate = TRUE)
}
\arguments{
\item{reg}{regulon object - see output of viper::aracne2regulon}

\item{annotate}{logical, whether to add a column on target type in ouptu}
}
\value{
a tibble with targets and metrics (MOR/likelihood)
}
\description{
transform a regulon object to a data frame / tibble
}
