% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{replace_missings}
\alias{replace_missings}
\title{Replace missing values (typically in a vector or matrix)}
\usage{
replace_missings(x, replacement = 0)
}
\arguments{
\item{x}{Vector (or matrix) with NA values somewhere}

\item{replacement}{Defaults to zero for numeric data, but can be anything really}
}
\value{
new matrix with randomized columns or rows, respectively
}
\description{
Replace missing values (typically in a vector or matrix)
}
