% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{rescale01}
\alias{rescale01}
\title{Rescale a numeric vector to range between 0 and 1}
\usage{
rescale01(x, na_rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{na_rm}{logical, whether to remove NA values in call to range (TRUE)}
}
\value{
rescaled numeric vector (minimum is 0, maximum is 1)
}
\description{
Rescale a numeric vector to range between 0 and 1
}
