% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixOp.R
\name{rowTstat}
\alias{rowTstat}
\title{Calculate row-wise Welch’s t-statistic}
\usage{
rowTstat(X, cl)
}
\arguments{
\item{expmat}{numerical matrix, typically a gene expression matrix}

\item{classes}{integer vector indicating class assignment (0 or 1)}
}
\value{
numeric vector with Welch's t-statistics for each row
}
\description{
Calculate row-wise Welch’s t-statistic
}
