% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixOp.R
\name{rowVar}
\alias{rowVar}
\title{rowVar}
\usage{
rowVar(A, removeNA = TRUE)
}
\arguments{
\item{A}{a numeric matrix}

\item{removeNA}{should NA values be removed from rowSums/rowMeans ?}
}
\value{
vector of variance per row
}
\description{
This function replaces the apply(A, 1, var) call
}
