% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{specific_n}
\alias{specific_n}
\title{For a signature matrix containing conditions/comparisons in its columns and typically gene names
in its rows, this function will find the most specific N genes for a given condition. It will do so by
computing for each gene the difference between the individual value for in a given condition and the
maximum value among the other conditions.}
\usage{
specific_n(mat, nn = 50, verbose = TRUE)
}
\arguments{
\item{mat}{numeric matrix}

\item{nn}{integer indicating how many genes to isolate for each condition/comparison (default = 50)}

\item{verbose}{logical indicating whether to message information on the process}
}
\value{
a data frame with gene names for from the rownames of the input matrix
}
\description{
For a signature matrix containing conditions/comparisons in its columns and typically gene names
in its rows, this function will find the most specific N genes for a given condition. It will do so by
computing for each gene the difference between the individual value for in a given condition and the
maximum value among the other conditions.
}
