% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{stouffer_z}
\alias{stouffer_z}
\title{Stouffer integration of Z scores}
\usage{
stouffer_z(x, weights = NULL)
}
\arguments{
\item{x}{a vector of Z scores}
}
\value{
Z an integrated Z score
}
\description{
This function integrates Z-scores (i.e. NES) by the Stouffer method
}
\examples{
zs <-c(1,3,5,2,3)
stouffer(zs)
}
