% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{topN_mat}
\alias{topN_mat}
\title{For a numeric matrix containing samples or conditions in its columns and typically genes
in its rows, this function will find the top N genes for a given condition. You can choose to retrieve
both genes with highest and lowest values, respectively or either one of them.}
\usage{
topN_mat(mat, direction = c("both", "up", "down"), nn = 25, verbose = TRUE)
}
\arguments{
\item{mat}{numeric matrix}

\item{direction}{character, which tail should be returned, defaults to both up and down}

\item{nn}{integer indicating how many genes to isolate for each tail (default = 25)}
}
\value{
a data frame with gene names for from the rownames of the input matrix
}
\description{
For a numeric matrix containing samples or conditions in its columns and typically genes
in its rows, this function will find the top N genes for a given condition. You can choose to retrieve
both genes with highest and lowest values, respectively or either one of them.
}
