% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{tpm}
\alias{tpm}
\title{tpm}
\usage{
tpm(
  counts,
  log = TRUE,
  offset = 0.25,
  type = c("Symbol", "Entrez"),
  species = c("Human", "Mouse"),
  filterZero = TRUE
)
}
\arguments{
\item{counts}{Numeric matrix of \strong{human} raw read counts per gene}

\item{log}{Logical, whether the data should be log2 transformed}

\item{offset}{Number to offset zero values when applying log transformation}

\item{type}{Character to indicate the gene identifier, defaults to gene symbol}

\item{species}{Character to indicate the genes' species origin, defaults to human}
}
\value{
Numeric matrix of TPM data
}
\description{
This function computes tpm for human exonic read counts
}
