% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper-helpers.R
\name{vpres2regulon}
\alias{vpres2regulon}
\title{Extract the top up and down regulated MR per sample}
\usage{
vpres2regulon(vpres, nn = 25, direction = c("both", "up", "down"),
  make_regulon = FALSE, reverse = FALSE, verbose = TRUE)
}
\arguments{
\item{vpres}{numeric matrix of NES as returned by the viper function}

\item{nn}{integer of number of top MR to consider for each regulon}

\item{direction}{character, which tail should be returned, defaults to both up and down}

\item{make_regulon}{logical, whether to set class of return object to "regulon"}

\item{reverse, }{logical, whether to reverse the top MR - useful for drug screening results and downstream applications}

\item{verbose}{logical, whether to message progress report}
}
\value{
list object with top dysregulated MR per sample
}
\description{
Extract the top up and down regulated MR per sample
}
