% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viper-helpers.R
\name{vpsig_limma}
\alias{vpsig_limma}
\title{viperSignature and nullmodel using limma}
\usage{
vpsig_limma(eset, ref, per = 1000, keep_all = FALSE, seed = 42,
  verbose = TRUE)
}
\arguments{
\item{eset}{normalized gene expression matrix with samples of interest}

\item{ref}{normalized gene expression matrix with reference samples}

\item{per}{integer, number of overall permutations}

\item{keep_all}{logical, whether to return signature for all samples (incl. reference samples)}

\item{verbose}{logical, whether to report on progress or not}
}
\description{
viperSignature and nullmodel using limma
}
