% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{write_matrix}
\alias{write_matrix}
\title{Convenience function for writing gene expression to file}
\usage{
write_matrix(eset, path)
}
\arguments{
\item{eset}{Numeric gene expression matrix with gene identifiers as row names}

\item{path}{character string indicating path and name of file}
}
\description{
Convenience function for writing gene expression to file
}
