% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{basic_rank_norm}
\alias{basic_rank_norm}
\title{Basic rank normalization of a gene expression matrix}
\usage{
basic_rank_norm(counts, tpm = TRUE)
}
\arguments{
\item{counts}{Matrix containing raw read counts (generally).}

\item{tpm}{logical, adjust for gene length before rank transformation}
}
\value{
basic rank normalized matrix.
}
\description{
Basic rank normalization of a gene expression matrix
}
