% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{gs_ovlp}
\alias{gs_ovlp}
\title{Order the columns of a heatmap within the constraints of a factor
Generally, a left to right gradient will be achieved where column means will
increase from left to right.}
\usage{
gs_ovlp(genelist, genesets)
}
\arguments{
\item{genelist}{character vector of gene identifiers}

\item{genesets}{named list of character vectors}
}
\value{
a tibble of genes and gene sets they appear in
}
\description{
Order the columns of a heatmap within the constraints of a factor
Generally, a left to right gradient will be achieved where column means will
increase from left to right.
}
