% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_signatures_plot.R
\name{plot_OneGs_MultSig}
\alias{plot_OneGs_MultSig}
\title{Plot ranks of a given gene set across several gene expression profiles from a numeric matrix.}
\usage{
plot_OneGs_MultSig(
  sigmat,
  geneset,
  color = "black",
  signatureNames = NULL,
  gs_label = NULL,
  column_col = NULL,
  scale_alphas = TRUE,
  relative_sigs = FALSE,
  ...
)
}
\arguments{
\item{sigmat}{gene expression matrix, usually of raw values such as read counts, TPM or FPKM}

\item{geneset}{a character vector representing one gene set}

\item{color}{single character string indicating a color for the bar lines}

\item{signatureNames}{Character vector of length 2, specifying the extremes of the signature}

\item{gs_label}{single character string to label the gene set under investigation}

\item{column_col}{vector of color ids (character, hex, etc) that must match the number of columns in the signature matrix.}

\item{relative_sigs}{logical, whether signatures are considered relative, e.g. t-stats or logFC. Will carry out fgsea.}

\item{...}{given for compatibility, particularly for the adjustment of cex for various text labels}

\item{scale_alpha}{logical, whether to scale alpha values for bars, defaults to TRUE}
}
\value{
Nothing, a plot is generated in the default output device
}
\description{
This function generates a plot of the member genes of a given gene set across multiple signatures.
Its first intention is to illustrate the so called 'percent rank' of a given gene from a given
gene set.
}
\examples{
sigmat <- matrix(rnorm(1e5), ncol = 10, nrow = 1e4)
rownames(sigmat) <- paste0('Gene', seq(nrow(sigmat)))
colnames(sigmat) <- paste0('Sample', seq(ncol(sigmat)))
set.seed(42)
gs <- sample(rownames(sigmat), size = 100)
plot_OneGs_MultSig(sigmat = sigmat, geneset = gs)
}
